% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-call.R
\name{.fec_req_perform}
\alias{.fec_req_perform}
\title{Choose and apply pagination strategy}
\usage{
.fec_req_perform(req, pagination, per_page, max_results, max_reqs, call)
}
\arguments{
\item{req}{The request object to modify.}

\item{pagination}{The pagination scheme to use. Currently either "none" (the
default) or "basic" (a scheme that uses \code{per_page} and returned \code{pages}
information). If an endpoint has a \code{per_page} argument, use "basic".}

\item{per_page}{The number of results to return per page.}

\item{max_results}{The maximum number of results to return. Note that
slightly more results may be returned if \code{max_results} is not evenly
divisible by 100.}

\item{max_reqs}{The maximum number of separate requests to perform. Passed to
the max_reqs argument of \code{\link[httr2:req_perform_iterative]{httr2::req_perform_iterative()}} when \code{next_req}
is supplied. The default \code{2} should likely be changed to \code{Inf} after you
validate the function.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
A list of \code{\link[httr2:response]{httr2::response()}} objects, one for each request
performed.
}
\description{
Choose and apply pagination strategy
}
\keyword{internal}
