% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/030-responses.R
\name{.fec_response_parser}
\alias{.fec_response_parser}
\title{Basic stop-gap parsing}
\usage{
.fec_response_parser(resp)
}
\arguments{
\item{resp}{A response object.}
}
\value{
A tibble of results.
}
\description{
Basic stop-gap parsing
}
\keyword{internal}
