% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-call.R
\name{fec_call_api}
\alias{fec_call_api}
\title{Call the OpenFEC API}
\usage{
fec_call_api(
  path,
  query = list(),
  body = NULL,
  method = NULL,
  pagination = c("none", "basic"),
  max_results = Inf,
  max_reqs = Inf,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY"),
  call = rlang::caller_env()
)
}
\arguments{
\item{path}{The route to an API endpoint. Optionally, a list with the path
plus variables to \code{\link[glue:glue]{glue::glue()}} into the path.}

\item{query}{An optional list of parameters to pass in the query portion of
the request.}

\item{body}{An object to use as the body of the request. If any component of
the body is a path, pass it through \code{\link[fs:path]{fs::path()}} or otherwise give it the
class "fs_path" to indicate that it is a path.}

\item{method}{If the method is something other than GET or POST, supply it.
Case is ignored.}

\item{pagination}{The pagination scheme to use. Currently either "none" (the
default) or "basic" (a scheme that uses \code{per_page} and returned \code{pages}
information). If an endpoint has a \code{per_page} argument, use "basic".}

\item{max_results}{The maximum number of results to return.}

\item{max_reqs}{The maximum number of separate requests to perform. Passed to
the max_reqs argument of \code{\link[httr2:req_perform_iterative]{httr2::req_perform_iterative()}} when \code{next_req}
is supplied. The default \code{2} should likely be changed to \code{Inf} after you
validate the function.}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
A tibble with the results of the API call.
}
\description{
Generate a request to an OpenFEC endpoint.
}
