% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-audit.R
\name{fec_get_audit_case}
\alias{fec_get_audit_case}
\title{Get audit case}
\usage{
fec_get_audit_case(
  audit_case_id,
  cycle,
  sub_category_id,
  sort_nulls_last,
  sort_hide_null,
  min_election_cycle,
  audit_id,
  q,
  per_page,
  max_election_cycle,
  candidate_id,
  committee_type,
  qq,
  page,
  committee_id,
  committee_designation,
  primary_category_id,
  sort_null_only,
  sort,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{audit_case_id}{Primary/foreign key for audit tables}

\item{cycle}{Filter records to only those that are applicable to a given two-year period. This cycle follows the traditional House election cycle and subdivides the presidential and Senate elections into comparable two-year blocks. The cycle begins with an odd year and is named for its ending, even year.}

\item{sub_category_id}{The finding id of an audit. Finding are a category of broader issues. Each category has an unique ID.}

\item{sort_nulls_last}{Toggle that sorts null values last}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{min_election_cycle}{Filter records to only those that are applicable to a given two-year period. This cycle follows the traditional House election cycle and subdivides the presidential and Senate elections into comparable two-year blocks. The cycle begins with an odd year and is named for its ending, even year.}

\item{audit_id}{The audit issue. Each subcategory has an unique ID}

\item{q}{The name of the committee. If a committee changes its name, the most recent name will be shown. Committee names are not unique. Use committee_id for looking up records.}

\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{max_election_cycle}{Filter records to only those that are applicable to a given two-year period. This cycle follows the traditional House election cycle and subdivides the presidential and Senate elections into comparable two-year blocks. The cycle begins with an odd year and is named for its ending, even year.}

\item{candidate_id}{A unique identifier assigned to each candidate registered with the FEC. If a person runs for several offices, that person will have separate candidate IDs for each office.}

\item{committee_type}{The one-letter type code of the organization: - C communication cost - D delegate - E electioneering communication - H House - I independent expenditure filer (not a committee) - N PAC - nonqualified - O independent expenditure-only (super PACs) - P presidential - Q PAC - qualified - S Senate - U single candidate independent expenditure - V PAC with non-contribution account, nonqualified - W PAC with non-contribution account, qualified - X party, nonqualified - Y party, qualified - Z national party non-federal account}

\item{qq}{Name of candidate running for office}

\item{page}{For paginating through results, starting at page 1}

\item{committee_id}{A unique identifier assigned to each committee or filer registered with the FEC. In general committee id's begin with the letter C which is followed by eight digits.}

\item{committee_designation}{Type of committee: - H or S - Congressional - P - Presidential - X or Y or Z - Party - N or Q - PAC - I - Independent expenditure - O - Super PAC}

\item{primary_category_id}{Audit category ID (table PK)}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{sort}{Provide a field to sort by. Use \code{-} for descending order. ex: \code{-case_no}}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
This endpoint contains Final Audit Reports approved by the Commission since inception. The search can be based on information about the audited committee (Name, FEC ID Number, Type, Election Cycle) or the issues covered in the report.
}
