% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-audit.R
\name{fec_get_audit_primary_category}
\alias{fec_get_audit_primary_category}
\title{Get audit primary category}
\usage{
fec_get_audit_primary_category(
  sort_nulls_last,
  page,
  primary_category_name,
  sort_hide_null,
  primary_category_id,
  sort_null_only,
  per_page,
  sort,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{sort_nulls_last}{Toggle that sorts null values last}

\item{page}{For paginating through results, starting at page 1}

\item{primary_category_name}{Primary Audit Category - No Findings or Issues/Not a Committee - Net Outstanding Campaign/Convention Expenditures/Obligations - Payments/Disgorgements - Allocation Issues - Prohibited Contributions - Disclosure - Recordkeeping - Repayment to US Treasury - Other - Misstatement of Financial Activity - Excessive Contributions - Failure to File Reports/Schedules/Notices - Loans - Referred Findings Not Listed}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{primary_category_id}{Audit category ID (table PK)}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{sort}{Provide a field to sort by. Use \code{-} for descending order.}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
This lists the options for the primary categories available in the /audit-search/ endpoint.
}
