% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-dates.R
\name{fec_get_calendar_dates_export}
\alias{fec_get_calendar_dates_export}
\title{Get calendar dates export}
\usage{
fec_get_calendar_dates_export(
  renderer,
  sort_nulls_last,
  page,
  min_start_date,
  calendar_category_id,
  sort_hide_null,
  min_end_date,
  event_id,
  sort_null_only,
  per_page,
  description,
  sort,
  max_end_date,
  summary,
  max_start_date,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{renderer}{BKTODO: No description provided.}

\item{sort_nulls_last}{Toggle that sorts null values last}

\item{page}{For paginating through results, starting at page 1}

\item{min_start_date}{The minimum start date.(MM/DD/YYYY or YYYY-MM-DD)}

\item{calendar_category_id}{Each type of event has a calendar category with an integer id. Options are: Open Meetings: 32, Executive Sessions: 39, Public Hearings: 40, Conferences: 33, Roundtables: 34, Election Dates: 36, Federal Holidays: 37, FEA Periods: 38, Commission Meetings: 20, Reporting Deadlines: 21, Conferences and Outreach: 22, AOs and Rules: 23, Other: 24, Quarterly: 25, Monthly: 26, Pre and Post-Elections: 27, EC Periods:28, and IE Periods: 29}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{min_end_date}{The minimum end date.(MM/DD/YYYY or YYYY-MM-DD)}

\item{event_id}{An unique ID for an event. Useful for downloading a single event to your calendar. This ID is not a permanent, persistent ID.}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{description}{Brief description of event}

\item{sort}{Provide a field to sort by. Use \code{-} for descending order.}

\item{max_end_date}{The maximum end date.(MM/DD/YYYY or YYYY-MM-DD)}

\item{summary}{Longer description of event}

\item{max_start_date}{The maximum start date.(MM/DD/YYYY or YYYY-MM-DD)}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
Returns CSV or ICS for downloading directly into calendar applications like Google, Outlook or other applications. Combines the election and reporting dates with Commission meetings, conferences, outreach, Advisory Opinions, rules, litigation dates and other events into one calendar. State filtering now applies to elections, reports and reporting periods. Presidential pre-primary report due dates are not shown on even years. Filers generally opt to file monthly rather than submit over 50 pre-primary election reports. All reporting deadlines are available at /reporting-dates/ for reference. This is \href{https://github.com/fecgov/openFEC/blob/develop/data/migrations/V40__omnibus_dates.sql}{the sql function} that creates the calendar.
}
