% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-committee.R
\name{fec_get_candidate_candidate_id_committees_history_cycle}
\alias{fec_get_candidate_candidate_id_committees_history_cycle}
\title{Get candidate candidate id committees history cycle}
\usage{
fec_get_candidate_candidate_id_committees_history_cycle(
  designation,
  sort_nulls_last,
  page,
  sort_hide_null,
  election_full,
  sort_null_only,
  per_page,
  sort,
  candidate_id,
  cycle,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{designation}{The one-letter designation code of the organization: - A authorized by a candidate - J joint fundraising committee - P principal campaign committee of a candidate - U unauthorized - B lobbyist/registrant PAC - D leadership PAC}

\item{sort_nulls_last}{Toggle that sorts null values last}

\item{page}{For paginating through results, starting at page 1}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{election_full}{\code{True} indicates that full election period of a candidate. \code{False} indicates that two year election cycle.}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{sort}{Provide a field to sort by. Use \code{-} for descending order.}

\item{candidate_id}{A unique identifier assigned to each candidate registered with the FEC. If a person runs for several offices, that person will have separate candidate IDs for each office.}

\item{cycle}{A two year election cycle that the committee was active- (after original registration date but before expiration date in Form 1s) The cycle begins with an odd year and is named for its ending, even year.}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
Explore a filer's characteristics over time. This can be particularly useful if the committees change treasurers, designation, or \code{committee_type}.
}
