% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-candidate.R
\name{fec_get_candidate_candidate_id_history}
\alias{fec_get_candidate_candidate_id_history}
\title{Get candidate candidate id history}
\usage{
fec_get_candidate_candidate_id_history(
  sort_nulls_last,
  per_page,
  sort_null_only,
  sort_hide_null,
  sort,
  election_full,
  page,
  candidate_id,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{sort_nulls_last}{Toggle that sorts null values last}

\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{sort}{Provide a field to sort by. Use \code{-} for descending order.}

\item{election_full}{\code{True} indicates that full election period of a candidate. \code{False} indicates that two year election cycle.}

\item{page}{For paginating through results, starting at page 1}

\item{candidate_id}{A unique identifier assigned to each candidate registered with the FEC. If a person runs for several offices, that person will have separate candidate IDs for each office.}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
Find out a candidate's characteristics over time. This is particularly useful if the candidate runs for the same office in different districts or you want to know more about a candidate's previous races. This information is organized by \code{candidate_id}, so it won't help you find a candidate who ran for different offices over time; candidates get a new ID for each office.
}
