% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-committee.R
\name{fec_get_committee_committee_id}
\alias{fec_get_committee_committee_id}
\title{Get committee committee id}
\usage{
fec_get_committee_committee_id(
  designation,
  committee_type,
  page,
  filing_frequency,
  cycle,
  sort_hide_null,
  year,
  sort_nulls_last,
  sort_null_only,
  per_page,
  organization_type,
  sort,
  committee_id,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{designation}{The one-letter designation code of the organization: - A authorized by a candidate - J joint fundraising committee - P principal campaign committee of a candidate - U unauthorized - B lobbyist/registrant PAC - D leadership PAC}

\item{committee_type}{The one-letter type code of the organization: - C communication cost - D delegate - E electioneering communication - H House - I independent expenditure filer (not a committee) - N PAC - nonqualified - O independent expenditure-only (super PACs) - P presidential - Q PAC - qualified - S Senate - U single candidate independent expenditure - V PAC with non-contribution account, nonqualified - W PAC with non-contribution account, qualified - X party, nonqualified - Y party, qualified - Z national party non-federal account}

\item{page}{For paginating through results, starting at page 1}

\item{filing_frequency}{The one-letter code of the filing frequency: - A Administratively terminated - D Debt - M Monthly filer - Q Quarterly filer - T Terminated - W Waived}

\item{cycle}{A two year election cycle that the committee was active- (after original registration date but before expiration date in Form 1s) The cycle begins with an odd year and is named for its ending, even year.}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{year}{A year that the committee was active— (after original registration date or filing but before expiration date)}

\item{sort_nulls_last}{Toggle that sorts null values last}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{organization_type}{The one-letter code for the kind for organization: - C corporation - L labor organization - M membership organization - T trade association - V cooperative - W corporation without capital stock}

\item{sort}{Provide a field to sort by. Use \code{-} for descending order.}

\item{committee_id}{A unique identifier assigned to each committee or filer registered with the FEC. In general committee id's begin with the letter C which is followed by eight digits.}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
This endpoint is useful for finding detailed information about a particular committee or filer. Use the \code{committee_id} to find the most recent information about the committee.
}
