% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-filings.R
\name{fec_get_committee_committee_id_filings}
\alias{fec_get_committee_committee_id_filings}
\title{Get committee committee id filings}
\usage{
fec_get_committee_committee_id_filings(
  office,
  file_number,
  form_category,
  cycle,
  district,
  document_type,
  sort_hide_null,
  state,
  amendment_indicator,
  sort_nulls_last,
  is_amended,
  filer_type,
  max_receipt_date,
  per_page,
  primary_general_indicator,
  request_type,
  report_year,
  form_type,
  committee_type,
  beginning_image_number,
  page,
  sort,
  min_receipt_date,
  sort_null_only,
  q_filer,
  report_type,
  party,
  most_recent,
  committee_id,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{office}{Federal office candidate runs for: H, S or P}

\item{file_number}{Filing ID number}

\item{form_category}{The forms filed are categorized based on the nature of the filing: - REPORT F3, F3X, F3P, F3L, F4, F5, F7, F13 - NOTICE F5, F24, F6, F9, F10, F11 - STATEMENT F1, F2 - OTHER F1M, F8, F99, F12, FRQ}

\item{cycle}{Filter records to only those that were applicable to a given two-year period.The cycle begins with an odd year and is named for its ending, even year.}

\item{district}{Two-digit US House distirict of the office the candidate is running for. Presidential, Senate and House at-large candidates will have District 00.}

\item{document_type}{The type of document for documents other than reports: - 2 24 Hour Contribution Notice - 4 48 Hour Contribution Notice - A Debt Settlement Statement - B Acknowledgment of Receipt of Debt Settlement Statement - C RFAI: Debt Settlement First Notice - D Commission Debt Settlement Review - E Commission Response TO Debt Settlement Request - F Administrative Termination - G Debt Settlement Plan Amendment - H Disavowal Notice - I Disavowal Response - J Conduit Report - K Termination Approval - L Repeat Non-Filer Notice - M Filing Frequency Change Notice - N Paper Amendment to Electronic Report - O Acknowledgment of Filing Frequency Change - S RFAI: Debt Settlement Second - T Miscellaneous Report TO FEC - V Repeat Violation Notice (441A OR 441B) - P Notice of Paper Filing - R F3L Filing Frequency Change Notice - Q Acknowledgment of F3L Filing Frequency Change - U Unregistered Committee Notice}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{state}{US state or territory where a candidate runs for office}

\item{amendment_indicator}{Amendent types: -N new -A amendment -T terminated -C consolidated -M multi-candidate -S secondary NULL might be new or amendment. If amendment indicator is null and the filings is the first or first in a chain treat it as if it was a new. If it is not the first or first in a chain then treat the filing as an amendment.}

\item{sort_nulls_last}{Toggle that sorts null values last}

\item{is_amended}{False indicates that a report is the most recent. True indicates that the report has been superseded by an amendment.}

\item{filer_type}{The method used to file with the FEC, either electronic or on paper.}

\item{max_receipt_date}{Selects all filings received before this date(MM/DD/YYYY or YYYY-MM-DD)}

\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{primary_general_indicator}{Primary, general or special election indicator.}

\item{request_type}{Requests for additional information (RFAIs) sent to filers. The request type is based on the type of document filed: - 1 Statement of Organization - 2 Report of Receipts and Expenditures (Form 3 and 3X) - 3 Second Notice - Reports - 4 Request for Additional Information - 5 Informational - Reports - 6 Second Notice - Statement of Organization - 7 Failure to File - 8 From Public Disclosure - 9 From Multi Candidate Status}

\item{report_year}{Forms with coverage date - year from the coverage ending date. Forms without coverage date - year from the receipt date.}

\item{form_type}{The form where the underlying data comes from, for example, Form 1 would appear as F1: - F1 Statement of Organization - F1M Notification of Multicandidate Status - F2 Statement of Candidacy - F3 Report of Receipts and Disbursements for an Authorized Committee - F3P Report of Receipts and Disbursements by an Authorized Committee of a Candidate for The Office of President or Vice President - F3L Report of Contributions Bundled by Lobbyists/Registrants and Lobbyist/Registrant PACs - F3X Report of Receipts and Disbursements for other than an Authorized Committee - F4 Report of Receipts and Disbursements for a Committee or Organization Supporting a Nomination Convention - F5 Report of Independent Expenditures Made and Contributions Received - F6 48 Hour Notice of Contributions/Loans Received - F7 Report of Communication Costs by Corporations and Membership Organizations - F8 Debt Settlement Plan - F9 24 Hour Notice of Disbursements for Electioneering Communications - F13 Report of Donations Accepted for Inaugural Committee - F99 Miscellaneous Text - FRQ Request for Additional Information}

\item{committee_type}{The one-letter type code of the organization: - C communication cost - D delegate - E electioneering communication - H House - I independent expenditure filer (not a committee) - N PAC - nonqualified - O independent expenditure-only (super PACs) - P presidential - Q PAC - qualified - S Senate - U single candidate independent expenditure - V PAC with non-contribution account, nonqualified - W PAC with non-contribution account, qualified - X party, nonqualified - Y party, qualified - Z national party non-federal account}

\item{beginning_image_number}{Unique identifier for the electronic or paper report. This number is used to construct PDF URLs to the original document.}

\item{page}{For paginating through results, starting at page 1}

\item{sort}{Provide a field to sort by. Use \code{-} for descending order. ex: \code{-case_no}}

\item{min_receipt_date}{Selects all filings received after this date(MM/DD/YYYY or YYYY-MM-DD)}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{q_filer}{Keyword search for filer name or ID}

\item{report_type}{Name of report where the underlying data comes from: - 10D Pre-Election - 10G Pre-General - 10P Pre-Primary - 10R Pre-Run-Off - 10S Pre-Special - 12C Pre-Convention - 12G Pre-General - 12P Pre-Primary - 12R Pre-Run-Off - 12S Pre-Special - 30D Post-Election - 30G Post-General - 30P Post-Primary - 30R Post-Run-Off - 30S Post-Special - 60D Post-Convention - M1 January Monthly - M10 October Monthly - M11 November Monthly - M12 December Monthly - M2 February Monthly - M3 March Monthly - M4 April Monthly - M5 May Monthly - M6 June Monthly - M7 July Monthly - M8 August Monthly - M9 September Monthly - MY Mid-Year Report - Q1 April Quarterly - Q2 July Quarterly - Q3 October Quarterly - TER Termination Report - YE Year-End - ADJ COMP ADJUST AMEND - CA COMPREHENSIVE AMEND - 90S Post Inaugural Supplement - 90D Post Inaugural - 48 48 Hour Notification - 24 24 Hour Notification - M7S July Monthly/Semi-Annual - MSA Monthly Semi-Annual (MY) - MYS Monthly Year End/Semi-Annual - Q2S July Quarterly/Semi-Annual - QSA Quarterly Semi-Annual (MY) - QYS Quarterly Year End/Semi-Annual - QYE Quarterly Semi-Annual (YE) - QMS Quarterly Mid-Year/ Semi-Annual - MSY Monthly Semi-Annual (YE)}

\item{party}{Three-letter code for the party affiliated with a candidate or committee. For example, DEM for Democratic Party and REP for Republican Party.}

\item{most_recent}{Report is either new or is the most-recently filed amendment}

\item{committee_id}{A unique identifier assigned to each committee or filer registered with the FEC. In general committee id's begin with the letter C which is followed by eight digits.}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
All official records and reports filed by or delivered to the FEC. Note: because the filings data includes many records, counts for large result sets are approximate; you will want to page through the records until no records are returned.
}
