% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-financial.R
\name{fec_get_committee_committee_id_totals}
\alias{fec_get_committee_committee_id_totals}
\title{Get committee committee id totals}
\usage{
fec_get_committee_committee_id_totals(
  cycle,
  sort_nulls_last,
  per_page,
  sort_null_only,
  sort_hide_null,
  sort,
  page,
  committee_id,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{cycle}{Filter records to only those that were applicable to a given two-year period.The cycle begins with an odd year and is named for its ending, even year.}

\item{sort_nulls_last}{Toggle that sorts null values last}

\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{sort}{Provide a field to sort by. Use \code{-} for descending order.}

\item{page}{For paginating through results, starting at page 1}

\item{committee_id}{A unique identifier assigned to each committee or filer registered with the FEC. In general committee id's begin with the letter C which is followed by eight digits.}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
This endpoint provides information about a committee's Form 3, Form 3X, or Form 3P financial reports, which are aggregated by two-year period. We refer to two-year periods as a \code{cycle}. The cycle is named after the even-numbered year and includes the year before it. To obtain totals from 2013 and 2014, you would use 2014. In odd-numbered years, the current cycle is the next year — for example, in 2015, the current cycle is 2016. For presidential and Senate candidates, multiple two-year cycles exist between elections.
}
