% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-committee.R
\name{fec_get_committees}
\alias{fec_get_committees}
\title{Get committees}
\usage{
fec_get_committees(
  designation,
  max_first_f1_date,
  min_first_f1_date,
  cycle,
  filing_frequency,
  sort_nulls_last,
  sort_hide_null,
  year,
  state,
  sponsor_candidate_id,
  q,
  per_page,
  min_last_f1_date,
  candidate_id,
  committee_type,
  min_first_file_date,
  page,
  sort,
  committee_id,
  sort_null_only,
  max_last_f1_date,
  treasurer_name,
  organization_type,
  max_first_file_date,
  party,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{designation}{The one-letter designation code of the organization: - A authorized by a candidate - J joint fundraising committee - P principal campaign committee of a candidate - U unauthorized - B lobbyist/registrant PAC - D leadership PAC}

\item{max_first_f1_date}{Filter for committees whose first Form 1 was received on or before this date.}

\item{min_first_f1_date}{Filter for committees whose first Form 1 was received on or after this date.}

\item{cycle}{A two year election cycle that the committee was active- (after original registration date but before expiration date in Form 1s) The cycle begins with an odd year and is named for its ending, even year.}

\item{filing_frequency}{The one-letter code of the filing frequency: - A Administratively terminated - D Debt - M Monthly filer - Q Quarterly filer - T Terminated - W Waived}

\item{sort_nulls_last}{Toggle that sorts null values last}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{year}{A year that the committee was active— (after original registration date or filing but before expiration date)}

\item{state}{US state or territory}

\item{sponsor_candidate_id}{A unique identifier assigned to each candidate registered with the FEC. If a person runs for several offices, that person will have separate candidate IDs for each office. This is a filter for Leadership PAC sponsor.}

\item{q}{The name of the committee. If a committee changes its name, the most recent name will be shown. Committee names are not unique. Use committee_id for looking up records.}

\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{min_last_f1_date}{Filter for committees whose latest Form 1 was received on or after this date.}

\item{candidate_id}{A unique identifier assigned to each candidate registered with the FEC. If a person runs for several offices, that person will have separate candidate IDs for each office.}

\item{committee_type}{The one-letter type code of the organization: - C communication cost - D delegate - E electioneering communication - H House - I independent expenditure filer (not a committee) - N PAC - nonqualified - O independent expenditure-only (super PACs) - P presidential - Q PAC - qualified - S Senate - U single candidate independent expenditure - V PAC with non-contribution account, nonqualified - W PAC with non-contribution account, qualified - X party, nonqualified - Y party, qualified - Z national party non-federal account}

\item{min_first_file_date}{Filter for committees whose first filing was received on or after this date.}

\item{page}{For paginating through results, starting at page 1}

\item{sort}{Provide a field to sort by. Use \code{-} for descending order.}

\item{committee_id}{A unique identifier assigned to each committee or filer registered with the FEC. In general committee id's begin with the letter C which is followed by eight digits.}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{max_last_f1_date}{Filter for committees whose latest Form 1 was received on or before this date.}

\item{treasurer_name}{Name of the Committee's treasurer. If multiple treasurers for the committee, the most recent treasurer will be shown.}

\item{organization_type}{The one-letter code for the kind for organization: - C corporation - L labor organization - M membership organization - T trade association - V cooperative - W corporation without capital stock}

\item{max_first_file_date}{Filter for committees whose first filing was received on or before this date.}

\item{party}{Three-letter code for the party affiliated with a candidate or committee. For example, DEM for Democratic Party and REP for Republican Party.}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
Fetch basic information about committees and filers. Use parameters to filter for particular characteristics.
}
