% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-communication_cost.R
\name{fec_get_communication_costs_totals_by_candidate}
\alias{fec_get_communication_costs_totals_by_candidate}
\title{Get communication costs totals by candidate}
\usage{
fec_get_communication_costs_totals_by_candidate(
  candidate_id,
  sort_nulls_last,
  page,
  cycle,
  sort_hide_null,
  election_full,
  sort_null_only,
  per_page,
  sort,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{candidate_id}{A unique identifier assigned to each candidate registered with the FEC. If a person runs for several offices, that person will have separate candidate IDs for each office.}

\item{sort_nulls_last}{Toggle that sorts null values last}

\item{page}{For paginating through results, starting at page 1}

\item{cycle}{Filter records to only those that were applicable to a given two-year period.The cycle begins with an odd year and is named for its ending, even year.}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{election_full}{\code{True} indicates that full election period of a candidate. \code{False} indicates that two year election cycle.}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{sort}{Provide a field to sort by. Use \code{-} for descending order.}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
Total communications costs aggregated across committees on supported or opposed candidates by cycle or candidate election year.
}
