% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-dates.R
\name{fec_get_election_dates}
\alias{fec_get_election_dates}
\title{Get election dates}
\usage{
fec_get_election_dates(
  max_update_date,
  sort_nulls_last,
  sort_hide_null,
  max_primary_general_date,
  min_update_date,
  max_create_date,
  per_page,
  min_election_date,
  election_party,
  election_type_id,
  page,
  min_create_date,
  min_primary_general_date,
  election_year,
  election_district,
  office_sought,
  sort_null_only,
  max_election_date,
  sort,
  election_state,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{max_update_date}{The maximum date this record was last updated.(MM/DD/YYYY or YYYY-MM-DD)}

\item{sort_nulls_last}{Toggle that sorts null values last}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{max_primary_general_date}{The maximum date of primary or general election.(MM/DD/YYYY or YYYY-MM-DD)}

\item{min_update_date}{The minimum date this record was last updated.(MM/DD/YYYY or YYYY-MM-DD)}

\item{max_create_date}{The maximum date this record was added to the system.(MM/DD/YYYY or YYYY-MM-DD)}

\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{min_election_date}{The minimum date of election.}

\item{election_party}{Party, if applicable.}

\item{election_type_id}{Election type id}

\item{page}{For paginating through results, starting at page 1}

\item{min_create_date}{The minimum date this record was added to the system.(MM/DD/YYYY or YYYY-MM-DD)}

\item{min_primary_general_date}{The minimum date of primary or general election.(MM/DD/YYYY or YYYY-MM-DD)}

\item{election_year}{Year of election}

\item{election_district}{House district of the office sought, if applicable.}

\item{office_sought}{House, Senate or presidential office.}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{max_election_date}{The maximum date of election.}

\item{sort}{Provide a field to sort by. Use \code{-} for descending order.}

\item{election_state}{State or territory of the office sought.}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
FEC election dates since 1995.
}
