% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-electioneering.R
\name{fec_get_electioneering}
\alias{fec_get_electioneering}
\title{Get electioneering}
\usage{
fec_get_electioneering(
  candidate_id,
  max_amount,
  sort_nulls_last,
  page,
  sort_hide_null,
  committee_id,
  min_date,
  min_amount,
  per_page,
  sort_null_only,
  description,
  sort,
  report_year,
  last_index,
  max_date,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{candidate_id}{A unique identifier assigned to each candidate registered with the FEC. If a person runs for several offices, that person will have separate candidate IDs for each office.}

\item{max_amount}{Filter for all amounts less than a value.}

\item{sort_nulls_last}{Toggle that sorts null values last}

\item{page}{For paginating through results, starting at page 1}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{committee_id}{A unique identifier assigned to each committee or filer registered with the FEC. In general committee id's begin with the letter C which is followed by eight digits.}

\item{min_date}{Minimum disbursement date}

\item{min_amount}{Filter for all amounts greater than a value.}

\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{description}{BKTODO: No description provided.}

\item{sort}{Provide a field to sort by. Use \code{-} for descending order.}

\item{report_year}{Forms with coverage date - year from the coverage ending date. Forms without coverage date - year from the receipt date.}

\item{last_index}{Index of last result from previous page}

\item{max_date}{Maximum disbursement date}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
An electioneering communication is any broadcast, cable or satellite communication that fulfills each of the following conditions: \emph{The communication refers to a clearly identified federal candidate.} \emph{The communication is publicly distributed by a television station, radio station, cable television system or satellite system for a fee.} \emph{The communication is distributed within 60 days prior to a general election or 30 days prior to a primary election to federal office.}
}
