% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-financial.R
\name{fec_get_elections_search}
\alias{fec_get_elections_search}
\title{Get elections search}
\usage{
fec_get_elections_search(
  office,
  cycle,
  district,
  zip,
  state,
  sort_hide_null,
  sort_nulls_last,
  page,
  sort_null_only,
  per_page,
  sort,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{office}{BKTODO: No description provided.}

\item{cycle}{Two-year election cycle in which a candidate runs for office. Calculated from Form 2. The cycle begins with an odd year and is named for its ending, even year. This cycle follows the traditional house election cycle and subdivides the presidential and Senate elections into comparable two-year blocks. To retrieve data for the entire four years of a presidential term or six years of a senatorial term, you will need the \code{election_full} flag.}

\item{district}{Two-digit US House distirict of the office the candidate is running for. Presidential, Senate and House at-large candidates will have District 00.}

\item{zip}{Zip code}

\item{state}{US state or territory where a candidate runs for office}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{sort_nulls_last}{Toggle that sorts null values last}

\item{page}{For paginating through results, starting at page 1}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{sort}{Provide a field to sort by. Use \code{-} for descending order. ex: \code{-case_no}}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
List elections by cycle, office, state, and district.
}
