% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-financial.R
\name{fec_get_elections_summary}
\alias{fec_get_elections_summary}
\title{Get elections summary}
\usage{
fec_get_elections_summary(
  office,
  cycle,
  district,
  state,
  election_full,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{office}{Federal office candidate runs for: H, S or P}

\item{cycle}{Two-year election cycle in which a candidate runs for office. Calculated from Form 2. The cycle begins with an odd year and is named for its ending, even year. This cycle follows the traditional house election cycle and subdivides the presidential and Senate elections into comparable two-year blocks. To retrieve data for the entire four years of a presidential term or six years of a senatorial term, you will need the \code{election_full} flag.}

\item{district}{Two-digit US House distirict of the office the candidate is running for. Presidential, Senate and House at-large candidates will have District 00.}

\item{state}{US state or territory where a candidate runs for office}

\item{election_full}{\code{True} indicates that full election period of a candidate. \code{False} indicates that two year election cycle.}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
List elections by cycle, office, state, and district.
}
