% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-audit.R
\name{fec_get_names_audit_committees}
\alias{fec_get_names_audit_committees}
\title{Get names audit committees}
\usage{
fec_get_names_audit_committees(
  q,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{q}{Name (candidate or committee) to search for}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
Search for candidates or committees by name. If you're looking for information on a particular person or group, using a name to find the \code{candidate_id} or \code{committee_id} on this endpoint can be a helpful first step.
}
