% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-presidential.R
\name{fec_get_presidential_contributions_by_candidate}
\alias{fec_get_presidential_contributions_by_candidate}
\title{Get presidential contributions by candidate}
\usage{
fec_get_presidential_contributions_by_candidate(
  contributor_state,
  sort_nulls_last,
  page,
  sort_hide_null,
  election_year,
  sort_null_only,
  per_page,
  sort,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{contributor_state}{State of contributor}

\item{sort_nulls_last}{Toggle that sorts null values last}

\item{page}{For paginating through results, starting at page 1}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{election_year}{Year of election}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{sort}{Provide a field to sort by. Use \code{-} for descending order.}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
Net receipts per candidate. Filter with \code{contributor_state='US'} for national totals
}
