% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-presidential.R
\name{fec_get_presidential_coverage_end_date}
\alias{fec_get_presidential_coverage_end_date}
\title{Get presidential coverage end date}
\usage{
fec_get_presidential_coverage_end_date(
  candidate_id,
  sort_nulls_last,
  page,
  sort_hide_null,
  election_year,
  sort_null_only,
  per_page,
  sort,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{candidate_id}{A unique identifier assigned to each candidate registered with the FEC. If a person runs for several offices, that person will have separate candidate IDs for each office. -P00000001 All candidates -P00000002 Democrasts -P00000003 Republicans}

\item{sort_nulls_last}{Toggle that sorts null values last}

\item{page}{For paginating through results, starting at page 1}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{election_year}{Year of election}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{sort}{Provide a field to sort by. Use \code{-} for descending order.}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
Coverage end date per candidate. Filter by candidate_id and/or election_year
}
