% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-filer_resources.R
\name{fec_get_rad_analyst}
\alias{fec_get_rad_analyst}
\title{Get rad analyst}
\usage{
fec_get_rad_analyst(
  min_assignment_update_date,
  max_assignment_update_date,
  analyst_short_id,
  email,
  page,
  sort_nulls_last,
  title,
  sort_hide_null,
  committee_id,
  name,
  telephone_ext,
  analyst_id,
  sort_null_only,
  per_page,
  sort,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{min_assignment_update_date}{Filter results for assignment updates made after this date}

\item{max_assignment_update_date}{Filter results for assignment updates made before this date}

\item{analyst_short_id}{Short ID of RAD analyst}

\item{email}{Email of RAD analyst}

\item{page}{For paginating through results, starting at page 1}

\item{sort_nulls_last}{Toggle that sorts null values last}

\item{title}{Title of RAD analyst}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{committee_id}{A unique identifier assigned to each committee or filer registered with the FEC. In general committee id's begin with the letter C which is followed by eight digits.}

\item{name}{Name of RAD analyst}

\item{telephone_ext}{Telephone extension of RAD analyst}

\item{analyst_id}{ID of RAD analyst}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{sort}{Provide a field to sort by. Use \code{-} for descending order.}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
Use this endpoint to look up the RAD Analyst for a committee. The mission of the Reports Analysis Division (RAD) is to ensure that campaigns and political committees file timely and accurate reports that fully disclose their financial activities. RAD is responsible for reviewing statements and financial reports filed by political committees participating in federal elections, providing assistance and guidance to the committees to properly file their reports, and for taking appropriate action to ensure compliance with the Federal Election Campaign Act (FECA).
}
