% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-dates.R
\name{fec_get_reporting_dates}
\alias{fec_get_reporting_dates}
\title{Get reporting dates}
\usage{
fec_get_reporting_dates(
  max_update_date,
  page,
  max_due_date,
  min_create_date,
  min_due_date,
  sort_hide_null,
  sort_nulls_last,
  min_update_date,
  max_create_date,
  per_page,
  sort_null_only,
  sort,
  report_year,
  report_type,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{max_update_date}{The maximum date this record was last updated.(MM/DD/YYYY or YYYY-MM-DD)}

\item{page}{For paginating through results, starting at page 1}

\item{max_due_date}{The maximum date the report is due.(MM/DD/YYYY or YYYY-MM-DD)}

\item{min_create_date}{The minimum date this record was added to the system.(MM/DD/YYYY or YYYY-MM-DD)}

\item{min_due_date}{The minimum date the report is due.(MM/DD/YYYY or YYYY-MM-DD)}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{sort_nulls_last}{Toggle that sorts null values last}

\item{min_update_date}{The minimum date this record was last updated.(MM/DD/YYYY or YYYY-MM-DD)}

\item{max_create_date}{The maximum date this record was added to the system.(MM/DD/YYYY or YYYY-MM-DD)}

\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{sort}{Provide a field to sort by. Use \code{-} for descending order.}

\item{report_year}{Forms with coverage date - year from the coverage ending date. Forms without coverage date - year from the receipt date.}

\item{report_type}{Name of report where the underlying data comes from: - 10D Pre-Election - 10G Pre-General - 10P Pre-Primary - 10R Pre-Run-Off - 10S Pre-Special - 12C Pre-Convention - 12G Pre-General - 12P Pre-Primary - 12R Pre-Run-Off - 12S Pre-Special - 30D Post-Election - 30G Post-General - 30P Post-Primary - 30R Post-Run-Off - 30S Post-Special - 60D Post-Convention - M1 January Monthly - M10 October Monthly - M11 November Monthly - M12 December Monthly - M2 February Monthly - M3 March Monthly - M4 April Monthly - M5 May Monthly - M6 June Monthly - M7 July Monthly - M8 August Monthly - M9 September Monthly - MY Mid-Year Report - Q1 April Quarterly - Q2 July Quarterly - Q3 October Quarterly - TER Termination Report - YE Year-End - ADJ COMP ADJUST AMEND - CA COMPREHENSIVE AMEND - 90S Post Inaugural Supplement - 90D Post Inaugural - 48 48 Hour Notification - 24 24 Hour Notification - M7S July Monthly/Semi-Annual - MSA Monthly Semi-Annual (MY) - MYS Monthly Year End/Semi-Annual - Q2S July Quarterly/Semi-Annual - QSA Quarterly Semi-Annual (MY) - QYS Quarterly Year End/Semi-Annual - QYE Quarterly Semi-Annual (YE) - QMS Quarterly Mid-Year/ Semi-Annual - MSY Monthly Semi-Annual (YE)}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
FEC election dates since 1995.
}
