% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-receipts.R
\name{fec_get_schedules_schedule_a_by_state_totals}
\alias{fec_get_schedules_schedule_a_by_state_totals}
\title{Get schedules schedule a by state totals}
\usage{
fec_get_schedules_schedule_a_by_state_totals(
  committee_type,
  cycle,
  page,
  state,
  sort_hide_null,
  sort_nulls_last,
  sort_null_only,
  per_page,
  sort,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{committee_type}{The one-letter type code of the organization: - C communication cost - D delegate - E electioneering communication - H House - I independent expenditure filer (not a committee) - N PAC - nonqualified - O independent expenditure-only (super PACs) - P presidential - Q PAC - qualified - S Senate - U single candidate independent expenditure - V PAC with non-contribution account, nonqualified - W PAC with non-contribution account, qualified - X party, nonqualified - Y party, qualified - Z national party non-federal account - all All Committee Types - all_candidates All Candidate Committee Types (H, S, P) - all_pacs All PAC Committee Types (N, O, Q, V, W)}

\item{cycle}{Filter records to only those that were applicable to a given two-year period.The cycle begins with an odd year and is named for its ending, even year.}

\item{page}{For paginating through results, starting at page 1}

\item{state}{US state or territory}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{sort_nulls_last}{Toggle that sorts null values last}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{sort}{Provide a field to sort by. Use \code{-} for descending order.}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
This endpoint provides itemized individual contributions received by a committee, aggregated by contributor’s state, committee type and cycle. If you are interested in our “is_individual” methodology, review the \href{https://www.fec.gov/campaign-finance-data/about-campaign-finance-data/methodology}{methodology page}. Unitemized individual contributions are not included.
}
