% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-receipts.R
\name{fec_get_schedules_schedule_a_by_zip}
\alias{fec_get_schedules_schedule_a_by_zip}
\title{Get schedules schedule a by zip}
\usage{
fec_get_schedules_schedule_a_by_zip(
  page,
  cycle,
  zip,
  sort_hide_null,
  state,
  committee_id,
  sort_nulls_last,
  sort_null_only,
  per_page,
  sort,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{page}{For paginating through results, starting at page 1}

\item{cycle}{Filter records to only those that were applicable to a given two-year period.The cycle begins with an odd year and is named for its ending, even year.}

\item{zip}{Zip code of contributor}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{state}{State of contributor}

\item{committee_id}{A unique identifier assigned to each committee or filer registered with the FEC. In general committee id's begin with the letter C which is followed by eight digits.}

\item{sort_nulls_last}{Toggle that sorts null values last}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{sort}{Provide a field to sort by. Use \code{-} for descending order.}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
This endpoint provides itemized individual contributions received by a committee, aggregated by the contributor’s ZIP code. If you are interested in our “is_individual” methodology, review the \href{https://www.fec.gov/campaign-finance-data/about-campaign-finance-data/methodology}{methodology page}. Unitemized individual contributions are not included.
}
