% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-receipts.R
\name{fec_get_schedules_schedule_a_efile}
\alias{fec_get_schedules_schedule_a_efile}
\title{Get schedules schedule a efile}
\usage{
fec_get_schedules_schedule_a_efile(
  max_amount,
  contributor_occupation,
  max_image_number,
  sort_nulls_last,
  sort_hide_null,
  line_number,
  per_page,
  contributor_city,
  contributor_state,
  page,
  sort,
  committee_id,
  min_date,
  contributor_employer,
  contributor_name,
  min_image_number,
  sort_null_only,
  image_number,
  min_amount,
  max_date,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{max_amount}{Filter for all amounts less than a value.}

\item{contributor_occupation}{Occupation of contributor, filers need to make an effort to gather this information}

\item{max_image_number}{Maxium image number of the page where the schedule item is reported}

\item{sort_nulls_last}{Toggle that sorts null values last}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{line_number}{Filter for form and line number using the following format: \code{FORM-LINENUMBER}. For example an argument such as \code{F3X-16} would filter down to all entries from form \code{F3X} line number \code{16}.}

\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{contributor_city}{City of contributor}

\item{contributor_state}{State of contributor}

\item{page}{For paginating through results, starting at page 1}

\item{sort}{Provide a field to sort by. Use \code{-} for descending order.}

\item{committee_id}{A unique identifier assigned to each committee or filer registered with the FEC. In general committee id's begin with the letter C which is followed by eight digits.}

\item{min_date}{Minimum date}

\item{contributor_employer}{Employer of contributor, filers need to make an effort to gather this information}

\item{contributor_name}{Name of contributor}

\item{min_image_number}{Minium image number of the page where the schedule item is reported}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{image_number}{An unique identifier for each page where the electronic or paper filing is reported.}

\item{min_amount}{Filter for all amounts greater than a value.}

\item{max_date}{Maximum date}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
Efiling endpoints provide real-time campaign finance data received from electronic filers. Efiling endpoints only contain the most recent four months of data and don't contain the processed and coded data that you can find on other endpoints.
}
