% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-disbursements.R
\name{fec_get_schedules_schedule_b}
\alias{fec_get_schedules_schedule_b}
\title{Get schedules schedule b}
\usage{
fec_get_schedules_schedule_b(
  disbursement_description,
  max_amount,
  max_image_number,
  image_number,
  sort_hide_null,
  last_disbursement_date,
  line_number,
  spender_committee_designation,
  last_disbursement_amount,
  per_page,
  last_index,
  spender_committee_type,
  sort,
  two_year_transaction_period,
  min_date,
  committee_id,
  disbursement_purpose_category,
  recipient_name,
  recipient_state,
  recipient_city,
  min_image_number,
  spender_committee_org_type,
  sort_null_only,
  recipient_committee_id,
  min_amount,
  max_date,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{disbursement_description}{Description of disbursement}

\item{max_amount}{Filter for all amounts less than a value.}

\item{max_image_number}{Maxium image number of the page where the schedule item is reported}

\item{image_number}{An unique identifier for each page where the electronic or paper filing is reported.}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{last_disbursement_date}{When sorting by \code{disbursement_date}, this is populated with the \code{disbursement_date} of the last result. However, you will need to pass the index of that last result to \code{last_index} to get the next page.}

\item{line_number}{Filter for form and line number using the following format: \code{FORM-LINENUMBER}. For example an argument such as \code{F3X-16} would filter down to all entries from form \code{F3X} line number \code{16}.}

\item{spender_committee_designation}{The one-letter designation code of the organization: - A authorized by a candidate - J joint fundraising committee - P principal campaign committee of a candidate - U unauthorized - B lobbyist/registrant PAC - D leadership PAC}

\item{last_disbursement_amount}{When sorting by \code{disbursement_amount}, this is populated with the \code{disbursement_amount} of the last result. However, you will need to pass the index of that last result to \code{last_index} to get the next page.}

\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{last_index}{Index of last result from previous page}

\item{spender_committee_type}{The one-letter type code of the organization: - C communication cost - D delegate - E electioneering communication - H House - I independent expenditure filer (not a committee) - N PAC - nonqualified - O independent expenditure-only (super PACs) - P presidential - Q PAC - qualified - S Senate - U single candidate independent expenditure - V PAC with non-contribution account, nonqualified - W PAC with non-contribution account, qualified - X party, nonqualified - Y party, qualified - Z national party non-federal account}

\item{sort}{Provide a field to sort by. Use \code{-} for descending order.}

\item{two_year_transaction_period}{This is a two-year period that is derived from the year a transaction took place in the Itemized Schedule A and Schedule B tables. In cases where we have the date of the transaction (contribution_receipt_date in schedules/schedule_a, disbursement_date in schedules/schedule_b) the two_year_transaction_period is named after the ending, even-numbered year. If we do not have the date of the transaction, we fall back to using the report year (report_year in both tables) instead, making the same cycle adjustment as necessary. If no transaction year is specified, the results default to the most current cycle.}

\item{min_date}{Minimum date}

\item{committee_id}{A unique identifier assigned to each committee or filer registered with the FEC. In general committee id's begin with the letter C which is followed by eight digits.}

\item{disbursement_purpose_category}{Disbursement purpose category}

\item{recipient_name}{Name of the entity receiving the disbursement}

\item{recipient_state}{State of recipient}

\item{recipient_city}{City of recipient}

\item{min_image_number}{Minium image number of the page where the schedule item is reported}

\item{spender_committee_org_type}{The one-letter code for the kind for organization: - C corporation - L labor organization - M membership organization - T trade association - V cooperative - W corporation without capital stock}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{recipient_committee_id}{The FEC identifier should be represented here if the contributor is registered with the FEC.}

\item{min_amount}{Filter for all amounts greater than a value.}

\item{max_date}{Maximum date}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
Schedule B filings describe itemized disbursements. This data explains how committees and other filers spend their money. These figures are reported as part of forms F3, F3X and F3P. The data is divided in two-year periods, called \code{two_year_transaction_period}, which is derived from the \code{report_year} submitted of the corresponding form. If no value is supplied, the results will default to the most recent two-year period that is named after the ending, even-numbered year. Due to the large quantity of Schedule B filings, this endpoint is not paginated by page number. Instead, you can request the next page of results by adding the values in the \code{last_indexes} object from \code{pagination} to the URL of your last request. For example, when sorting by \code{disbursement_date}, you might receive a page of results with the following pagination information: \verb{pagination: \{ pages: 965191, per_page: 20, count: 19303814, last_indexes: \{ last_index: &quot;230906248&quot;, last_disbursement_date: &quot;2014-07-04&quot; \} \}} To fetch the next page of sorted results, append \code{last_index=230906248} and \code{last_disbursement_date=2014-07-04} to the URL. We strongly advise paging through these results by using the sort indices (defaults to sort by disbursement date, e.g. \code{last_disbursement_date}), otherwise some resources may be unintentionally filtered out. This resource uses keyset pagination to improve query performance and these indices are required to properly page through this large dataset. Note: because the Schedule B data includes many records, counts for large result sets are approximate; you will want to page through the records until no records are returned.
}
