% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-disbursements.R
\name{fec_get_schedules_schedule_b_by_recipient_id}
\alias{fec_get_schedules_schedule_b_by_recipient_id}
\title{Get schedules schedule b by recipient id}
\usage{
fec_get_schedules_schedule_b_by_recipient_id(
  sort_nulls_last,
  page,
  cycle,
  sort_hide_null,
  committee_id,
  sort_null_only,
  per_page,
  recipient_id,
  sort,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{sort_nulls_last}{Toggle that sorts null values last}

\item{page}{For paginating through results, starting at page 1}

\item{cycle}{Filter records to only those that were applicable to a given two-year period.The cycle begins with an odd year and is named for its ending, even year.}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{committee_id}{A unique identifier assigned to each committee or filer registered with the FEC. In general committee id's begin with the letter C which is followed by eight digits.}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{recipient_id}{The FEC identifier should be represented here if the entity receiving the disbursement is registered with the FEC.}

\item{sort}{Provide a field to sort by. Use \code{-} for descending order.}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
Schedule B disbursements aggregated by recipient committee ID, if applicable. To avoid double counting, memoed items are not included.
}
