% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-disbursements.R
\name{fec_get_schedules_schedule_b_efile}
\alias{fec_get_schedules_schedule_b_efile}
\title{Get schedules schedule b efile}
\usage{
fec_get_schedules_schedule_b_efile(
  disbursement_description,
  max_amount,
  page,
  sort_nulls_last,
  sort,
  sort_hide_null,
  committee_id,
  min_date,
  recipient_state,
  recipient_city,
  sort_null_only,
  per_page,
  image_number,
  min_amount,
  max_date,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{disbursement_description}{Description of disbursement}

\item{max_amount}{Filter for all amounts less than a value.}

\item{page}{For paginating through results, starting at page 1}

\item{sort_nulls_last}{Toggle that sorts null values last}

\item{sort}{Provide a field to sort by. Use \code{-} for descending order.}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{committee_id}{A unique identifier assigned to each committee or filer registered with the FEC. In general committee id's begin with the letter C which is followed by eight digits.}

\item{min_date}{When sorting by \code{disbursement_date}, this is populated with the \code{disbursement_date} of the last result. However, you will need to pass the index of that last result to \code{last_index} to get the next page.}

\item{recipient_state}{State of recipient}

\item{recipient_city}{City of recipient}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{image_number}{An unique identifier for each page where the electronic or paper filing is reported.}

\item{min_amount}{Filter for all amounts less than a value.}

\item{max_date}{When sorting by \code{disbursement_date}, this is populated with the \code{disbursement_date} of the last result. However, you will need to pass the index of that last result to \code{last_index} to get the next page.}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
Efiling endpoints provide real-time campaign finance data received from electronic filers. Efiling endpoints only contain the most recent four months of data and don't contain the processed and coded data that you can find on other endpoints.
}
