% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-loans.R
\name{fec_get_schedules_schedule_c}
\alias{fec_get_schedules_schedule_c}
\title{Get schedules schedule c}
\usage{
fec_get_schedules_schedule_c(
  min_incurred_date,
  candidate_name,
  max_amount,
  max_image_number,
  sort_nulls_last,
  sort_hide_null,
  line_number,
  max_incurred_date,
  per_page,
  last_index,
  loan_source_name,
  page,
  sort,
  max_payment_to_date,
  min_payment_to_date,
  committee_id,
  min_image_number,
  sort_null_only,
  image_number,
  min_amount,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{min_incurred_date}{Minimum incurred date}

\item{candidate_name}{Name of candidate running for office}

\item{max_amount}{Filter for all amounts less than a value.}

\item{max_image_number}{Maxium image number of the page where the schedule item is reported}

\item{sort_nulls_last}{Toggle that sorts null values last}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{line_number}{Filter for form and line number using the following format: \code{FORM-LINENUMBER}. For example an argument such as \code{F3X-16} would filter down to all entries from form \code{F3X} line number \code{16}.}

\item{max_incurred_date}{Maximum incurred date}

\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{last_index}{Index of last result from previous page}

\item{loan_source_name}{Source of the loan (i.e., bank loan, brokerage account, credit card, home equity line of credit, other line of credit, or personal funds of the candidate}

\item{page}{For paginating through results, starting at page 1}

\item{sort}{Provide a field to sort by. Use \code{-} for descending order.}

\item{max_payment_to_date}{Maximum payment to date}

\item{min_payment_to_date}{Minimum payment to date}

\item{committee_id}{A unique identifier assigned to each committee or filer registered with the FEC. In general committee id's begin with the letter C which is followed by eight digits.}

\item{min_image_number}{Minium image number of the page where the schedule item is reported}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{image_number}{An unique identifier for each page where the electronic or paper filing is reported.}

\item{min_amount}{Filter for all amounts greater than a value.}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
Schedule C shows all loans, endorsements and loan guarantees a committee receives or makes. The committee continues to report the loan until it is repaid.
}
