% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-loans.R
\name{fec_get_schedules_schedule_c_sub_id}
\alias{fec_get_schedules_schedule_c_sub_id}
\title{Get schedules schedule c sub id}
\usage{
fec_get_schedules_schedule_c_sub_id(
  sort_nulls_last,
  per_page,
  sort_null_only,
  sort_hide_null,
  sort,
  page,
  sub_id,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{sort_nulls_last}{Toggle that sorts null values last}

\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{sort}{Provide a field to sort by. Use \code{-} for descending order.}

\item{page}{For paginating through results, starting at page 1}

\item{sub_id}{BKTODO: No description provided.}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
Schedule C shows all loans, endorsements and loan guarantees a committee receives or makes. The committee continues to report the loan until it is repaid.
}
