% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-debts.R
\name{fec_get_schedules_schedule_d}
\alias{fec_get_schedules_schedule_d}
\title{Get schedules schedule d}
\usage{
fec_get_schedules_schedule_d(
  creditor_debtor_name,
  max_image_number,
  sort_nulls_last,
  max_amount_outstanding_beginning,
  sort_hide_null,
  min_payment_period,
  max_amount_incurred,
  nature_of_debt,
  per_page,
  max_amount_outstanding_close,
  candidate_id,
  page,
  min_date,
  committee_id,
  min_amount_outstanding_close,
  max_payment_period,
  min_image_number,
  min_amount_incurred,
  sort_null_only,
  image_number,
  sort,
  min_amount_outstanding_beginning,
  max_date,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{creditor_debtor_name}{BKTODO: No description provided.}

\item{max_image_number}{Maxium image number of the page where the schedule item is reported}

\item{sort_nulls_last}{Toggle that sorts null values last}

\item{max_amount_outstanding_beginning}{BKTODO: No description provided.}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{min_payment_period}{BKTODO: No description provided.}

\item{max_amount_incurred}{BKTODO: No description provided.}

\item{nature_of_debt}{BKTODO: No description provided.}

\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{max_amount_outstanding_close}{BKTODO: No description provided.}

\item{candidate_id}{A unique identifier assigned to each candidate registered with the FEC. If a person runs for several offices, that person will have separate candidate IDs for each office.}

\item{page}{For paginating through results, starting at page 1}

\item{min_date}{Minimum load date}

\item{committee_id}{A unique identifier assigned to each committee or filer registered with the FEC. In general committee id's begin with the letter C which is followed by eight digits.}

\item{min_amount_outstanding_close}{BKTODO: No description provided.}

\item{max_payment_period}{BKTODO: No description provided.}

\item{min_image_number}{Minium image number of the page where the schedule item is reported}

\item{min_amount_incurred}{BKTODO: No description provided.}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{image_number}{An unique identifier for each page where the electronic or paper filing is reported.}

\item{sort}{Provide a field to sort by. Use \code{-} for descending order.}

\item{min_amount_outstanding_beginning}{BKTODO: No description provided.}

\item{max_date}{Maximum load date}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
Schedule D, it shows debts and obligations owed to or by the committee that are required to be disclosed.
}
