% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-independent_expenditures.R
\name{fec_get_schedules_schedule_e}
\alias{fec_get_schedules_schedule_e}
\title{Get schedules schedule e}
\usage{
fec_get_schedules_schedule_e(
  max_dissemination_date,
  sort_hide_null,
  payee_name,
  q_spender,
  per_page,
  last_index,
  min_dissemination_date,
  candidate_office_state,
  filing_form,
  sort_null_only,
  max_amount,
  max_image_number,
  min_filing_date,
  cycle,
  sort_nulls_last,
  support_oppose_indicator,
  candidate_office,
  is_notice,
  line_number,
  last_expenditure_amount,
  last_expenditure_date,
  candidate_id,
  last_office_total_ytd,
  last_support_oppose_indicator,
  min_date,
  candidate_party,
  committee_id,
  min_image_number,
  max_filing_date,
  candidate_office_district,
  min_amount,
  image_number,
  sort,
  max_date,
  most_recent,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{max_dissemination_date}{Selects all items distributed by this committee before this date}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{payee_name}{Name of the entity that received the payment.}

\item{q_spender}{Keyword search for spender name or ID}

\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{last_index}{Index of last result from previous page}

\item{min_dissemination_date}{Selects all items distributed by this committee after this date}

\item{candidate_office_state}{US state or territory}

\item{filing_form}{The form where the underlying data comes from, for example, Form 1 would appear as F1: - F1 Statement of Organization - F1M Notification of Multicandidate Status - F2 Statement of Candidacy - F3 Report of Receipts and Disbursements for an Authorized Committee - F3P Report of Receipts and Disbursements by an Authorized Committee of a Candidate for The Office of President or Vice President - F3L Report of Contributions Bundled by Lobbyists/Registrants and Lobbyist/Registrant PACs - F3X Report of Receipts and Disbursements for other than an Authorized Committee - F4 Report of Receipts and Disbursements for a Committee or Organization Supporting a Nomination Convention - F5 Report of Independent Expenditures Made and Contributions Received - F6 48 Hour Notice of Contributions/Loans Received - F7 Report of Communication Costs by Corporations and Membership Organizations - F8 Debt Settlement Plan - F9 24 Hour Notice of Disbursements for Electioneering Communications - F13 Report of Donations Accepted for Inaugural Committee - F99 Miscellaneous Text - FRQ Request for Additional Information}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{max_amount}{Filter for all amounts less than a value.}

\item{max_image_number}{Maxium image number of the page where the schedule item is reported}

\item{min_filing_date}{Selects all filings received after this date}

\item{cycle}{Filter records to only those that were applicable to a given two-year period.The cycle begins with an odd year and is named for its ending, even year.}

\item{sort_nulls_last}{Toggle that sorts null values last}

\item{support_oppose_indicator}{Explains if the money was spent in order to support or oppose a candidate or candidates. (Coded S or O for support or oppose.) This indicator applies to independent expenditures and communication costs.}

\item{candidate_office}{Federal office candidate runs for: H, S or P}

\item{is_notice}{Record filed as 24- or 48-hour notice.}

\item{line_number}{Filter for form and line number using the following format: \code{FORM-LINENUMBER}. For example an argument such as \code{F3X-16} would filter down to all entries from form \code{F3X} line number \code{16}.}

\item{last_expenditure_amount}{When sorting by \code{expenditure_amount}, this is populated with the \code{expenditure_amount} of the last result. However, you will need to pass the index of that last result to \code{last_index} to get the next page.}

\item{last_expenditure_date}{When sorting by \code{expenditure_date}, this is populated with the \code{expenditure_date} of the last result. However, you will need to pass the index of that last result to \code{last_index} to get the next page.}

\item{candidate_id}{A unique identifier assigned to each candidate registered with the FEC. If a person runs for several offices, that person will have separate candidate IDs for each office.}

\item{last_office_total_ytd}{When sorting by \code{office_total_ytd}, this is populated with the \code{office_total_ytd} of the last result. However, you will need to pass the index of that last result to \code{last_index} to get the next page.'}

\item{last_support_oppose_indicator}{When sorting by \code{support_oppose_indicator}, this is populated with the \code{support_oppose_indicator} of the last result. However, you will need to pass the index of that last result to \code{last_index} to get the next page.'}

\item{min_date}{Minimum date}

\item{candidate_party}{Three-letter code for the party affiliated with a candidate or committee. For example, DEM for Democratic Party and REP for Republican Party.}

\item{committee_id}{A unique identifier assigned to each committee or filer registered with the FEC. In general committee id's begin with the letter C which is followed by eight digits.}

\item{min_image_number}{Minium image number of the page where the schedule item is reported}

\item{max_filing_date}{Selects all filings received before this date}

\item{candidate_office_district}{Two-digit US House distirict of the office the candidate is running for. Presidential, Senate and House at-large candidates will have District 00.}

\item{min_amount}{Filter for all amounts greater than a value.}

\item{image_number}{An unique identifier for each page where the electronic or paper filing is reported.}

\item{sort}{Provide a field to sort by. Use \code{-} for descending order.}

\item{max_date}{Maximum date}

\item{most_recent}{The report associated with the transaction is either new or is the most-recently filed amendment. Undetermined version (\code{null}) is always included.}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
Schedule E covers the line item expenditures for independent expenditures. For example, if a super PAC bought ads on TV to oppose a federal candidate, each ad purchase would be recorded here with the expenditure amount, name and id of the candidate, and whether the ad supported or opposed the candidate. An independent expenditure is an expenditure for a communication "expressly advocating the election or defeat of a clearly identified candidate that is not made in cooperation, consultation, or concert with, or at the request or suggestion of, a candidate, a candidate’s authorized committee, or their agents, or a political party or its agents." Aggregates by candidate do not include 24 and 48 hour reports. This ensures we don't double count expenditures and the totals are more accurate. You can still find the information from 24 and 48 hour reports in \verb{/schedule/schedule_e/}. Due to the large quantity of Schedule E filings, this endpoint is not paginated by page number. Instead, you can request the next page of results by adding the values in the \code{last_indexes} object from \code{pagination} to the URL of your last request. For example, when sorting by \code{expenditure_amount}, you might receive a page of results with the following pagination information: \verb{&quot;pagination&quot;: \{ &quot;count&quot;: 152623, &quot;last_indexes&quot;: \{ &quot;last_index&quot;: &quot;3023037&quot;, &quot;last_expenditure_amount&quot;: -17348.5 \}, &quot;per_page&quot;: 20, &quot;pages&quot;: 7632 \} \}} To fetch the next page of sorted results, append \code{last_index=3023037} and \verb{last_expenditure_amount=} to the URL. We strongly advise paging through these results by using the sort indices (defaults to sort by disbursement date, e.g. \code{last_disbursement_date}), otherwise some resources may be unintentionally filtered out. This resource uses keyset pagination to improve query performance and these indices are required to properly page through this large dataset. Note: because the Schedule E data includes many records, counts for large result sets are approximate; you will want to page through the records until no records are returned.
}
