% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-independent_expenditures.R
\name{fec_get_schedules_schedule_e_by_candidate}
\alias{fec_get_schedules_schedule_e_by_candidate}
\title{Get schedules schedule e by candidate}
\usage{
fec_get_schedules_schedule_e_by_candidate(
  office,
  candidate_id,
  cycle,
  district,
  state,
  page,
  support_oppose,
  election_full,
  committee_id,
  sort_hide_null,
  sort_nulls_last,
  sort_null_only,
  per_page,
  sort,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{office}{Federal office candidate runs for: H, S or P}

\item{candidate_id}{A unique identifier assigned to each candidate registered with the FEC. If a person runs for several offices, that person will have separate candidate IDs for each office.}

\item{cycle}{Filter records to only those that were applicable to a given two-year period.The cycle begins with an odd year and is named for its ending, even year.}

\item{district}{Two-digit US House distirict of the office the candidate is running for. Presidential, Senate and House at-large candidates will have District 00.}

\item{state}{US state or territory where a candidate runs for office}

\item{page}{For paginating through results, starting at page 1}

\item{support_oppose}{Support or opposition}

\item{election_full}{\code{True} indicates that full election period of a candidate. \code{False} indicates that two year election cycle.}

\item{committee_id}{A unique identifier assigned to each committee or filer registered with the FEC. In general committee id's begin with the letter C which is followed by eight digits.}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{sort_nulls_last}{Toggle that sorts null values last}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{sort}{Provide a field to sort by. Use \code{-} for descending order.}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
Schedule E receipts aggregated by recipient candidate. To avoid double counting, memoed items are not included.
}
