% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-party_coordinated_expenditures.R
\name{fec_get_schedules_schedule_f}
\alias{fec_get_schedules_schedule_f}
\title{Get schedules schedule f}
\usage{
fec_get_schedules_schedule_f(
  max_amount,
  max_image_number,
  cycle,
  sort_nulls_last,
  sort_hide_null,
  line_number,
  payee_name,
  per_page,
  candidate_id,
  page,
  min_date,
  committee_id,
  min_amount,
  min_image_number,
  sort_null_only,
  image_number,
  sort,
  max_date,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{max_amount}{Filter for all amounts less than a value.}

\item{max_image_number}{Maxium image number of the page where the schedule item is reported}

\item{cycle}{Filter records to only those that were applicable to a given two-year period.The cycle begins with an odd year and is named for its ending, even year.}

\item{sort_nulls_last}{Toggle that sorts null values last}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{line_number}{Filter for form and line number using the following format: \code{FORM-LINENUMBER}. For example an argument such as \code{F3X-16} would filter down to all entries from form \code{F3X} line number \code{16}.}

\item{payee_name}{BKTODO: No description provided.}

\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{candidate_id}{A unique identifier assigned to each candidate registered with the FEC. If a person runs for several offices, that person will have separate candidate IDs for each office.}

\item{page}{For paginating through results, starting at page 1}

\item{min_date}{Minimum date}

\item{committee_id}{A unique identifier assigned to each committee or filer registered with the FEC. In general committee id's begin with the letter C which is followed by eight digits.}

\item{min_amount}{Filter for all amounts greater than a value.}

\item{min_image_number}{Minium image number of the page where the schedule item is reported}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{image_number}{An unique identifier for each page where the electronic or paper filing is reported.}

\item{sort}{Provide a field to sort by. Use \code{-} for descending order.}

\item{max_date}{Maximum date}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
Schedule F, it shows all special expenditures a national or state party committee makes in connection with the general election campaigns of federal candidates. These coordinated party expenditures do not count against the contribution limits but are subject to other limits, these limits are detailed in Chapter 7 of the FEC Campaign Guide for Political Party Committees.
}
