% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-party_coordinated_expenditures.R
\name{fec_get_schedules_schedule_f_sub_id}
\alias{fec_get_schedules_schedule_f_sub_id}
\title{Get schedules schedule f sub id}
\usage{
fec_get_schedules_schedule_f_sub_id(
  per_page,
  page,
  sub_id,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{page}{For paginating through results, starting at page 1}

\item{sub_id}{BKTODO: No description provided.}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
Schedule F, it shows all special expenditures a national or state party committee makes in connection with the general election campaigns of federal candidates. These coordinated party expenditures do not count against the contribution limits but are subject to other limits, these limits are detailed in Chapter 7 of the FEC Campaign Guide for Political Party Committees.
}
