% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-filer_resources.R
\name{fec_get_state_election_office}
\alias{fec_get_state_election_office}
\title{Get state election office}
\usage{
fec_get_state_election_office(
  sort_null_only,
  sort_nulls_last,
  per_page,
  state,
  sort_hide_null,
  sort,
  page,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{sort_nulls_last}{Toggle that sorts null values last}

\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{state}{Enter a state (Ex: AK, TX, VA etc..) to find the local election offices contact information.}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{sort}{Provide a field to sort by. Use \code{-} for descending order.}

\item{page}{For paginating through results, starting at page 1}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
State laws and procedures govern elections for state or local offices as well as how candidates appear on election ballots. Contact the appropriate state election office for more information.
}
