% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-financial.R
\name{fec_get_totals_by_entity}
\alias{fec_get_totals_by_entity}
\title{Get totals by entity}
\usage{
fec_get_totals_by_entity(
  cycle,
  sort_nulls_last,
  per_page,
  sort_null_only,
  sort_hide_null,
  sort,
  page,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{cycle}{Filter records to only those that were applicable to a given two-year period.The cycle begins with an odd year and is named for its ending, even year.}

\item{sort_nulls_last}{Toggle that sorts null values last}

\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{sort}{Provide a field to sort by. Use \code{-} for descending order.}

\item{page}{For paginating through results, starting at page 1}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
Provides cumulative receipt totals by entity type, over a two year cycle. Totals are adjusted to avoid double counting. This is \href{https://github.com/fecgov/openFEC/blob/develop/data/migrations/V41__large_aggregates.sql}{the sql} that creates these calculations.
}
