% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-financial.R
\name{fec_get_totals_entity_type}
\alias{fec_get_totals_entity_type}
\title{Get totals entity type}
\usage{
fec_get_totals_entity_type(
  max_first_f1_date,
  min_receipts,
  cycle,
  filing_frequency,
  max_receipts,
  sort_hide_null,
  sort_nulls_last,
  min_last_debts_owed_by_committee,
  max_last_cash_on_hand_end_period,
  treasurer_name,
  sponsor_candidate_id,
  per_page,
  max_disbursements,
  committee_state,
  committee_type,
  page,
  max_last_debts_owed_by_committee,
  committee_id,
  committee_designation,
  sort_null_only,
  min_first_f1_date,
  organization_type,
  sort,
  min_disbursements,
  min_last_cash_on_hand_end_period,
  entity_type,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)
}
\arguments{
\item{max_first_f1_date}{Filter for committees whose first Form 1 was received on or before this date.}

\item{min_receipts}{Filter for all amounts greater than a value.}

\item{cycle}{Filter records to only those that were applicable to a given two-year period.The cycle begins with an odd year and is named for its ending, even year.}

\item{filing_frequency}{The one-letter code of the filing frequency: - A Administratively terminated - D Debt - M Monthly filer - Q Quarterly filer - T Terminated - W Waived}

\item{max_receipts}{Filter for all amounts less than a value.}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{sort_nulls_last}{Toggle that sorts null values last}

\item{min_last_debts_owed_by_committee}{Filter for all amounts greater than a value.}

\item{max_last_cash_on_hand_end_period}{Filter for all amounts less than a value.}

\item{treasurer_name}{Name of the Committee's treasurer. If multiple treasurers for the committee, the most recent treasurer will be shown.}

\item{sponsor_candidate_id}{A unique identifier assigned to each candidate registered with the FEC. If a person runs for several offices, that person will have separate candidate IDs for each office. This is a filter for Leadership PAC sponsor.}

\item{per_page}{The number of results returned per page. Defaults to 20.}

\item{max_disbursements}{Filter for all amounts less than a value.}

\item{committee_state}{US state or territory}

\item{committee_type}{The one-letter type code of the organization: - C communication cost - D delegate - E electioneering communication - H House - I independent expenditure filer (not a committee) - N PAC - nonqualified - O independent expenditure-only (super PACs) - P presidential - Q PAC - qualified - S Senate - U single candidate independent expenditure - V PAC with non-contribution account, nonqualified - W PAC with non-contribution account, qualified - X party, nonqualified - Y party, qualified - Z national party non-federal account}

\item{page}{For paginating through results, starting at page 1}

\item{max_last_debts_owed_by_committee}{Filter for all amounts less than a value.}

\item{committee_id}{A unique identifier assigned to each committee or filer registered with the FEC. In general committee id's begin with the letter C which is followed by eight digits.}

\item{committee_designation}{The one-letter designation code of the organization: - A authorized by a candidate - J joint fundraising committee - P principal campaign committee of a candidate - U unauthorized - B lobbyist/registrant PAC - D leadership PAC}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{min_first_f1_date}{Filter for committees whose first Form 1 was received on or after this date.}

\item{organization_type}{The one-letter code for the kind for organization: - C corporation - L labor organization - M membership organization - T trade association - V cooperative - W corporation without capital stock}

\item{sort}{Provide a field to sort by. Use \code{-} for descending order.}

\item{min_disbursements}{Filter for all amounts greater than a value.}

\item{min_last_cash_on_hand_end_period}{Filter for all amounts greater than a value.}

\item{entity_type}{Committee groupings based on FEC filing form. Choose one of: \code{presidential}, \code{pac}, \code{party}, \code{pac-party}, \code{house-senate}, or \code{ie-only}}

\item{api_key}{An API key provided by the API provider. This key is not
clearly documented in the API description. Check the API documentation for
details.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
This endpoint provides information about a committee's Form 3, Form 3X, or Form 3P financial reports, which are aggregated by two-year period. We refer to two-year periods as a \code{cycle}. The cycle is named after the even-numbered year and includes the year before it. To obtain totals from 2013 and 2014, you would use 2014. In odd-numbered years, the current cycle is the next year — for example, in 2015, the current cycle is 2016. For presidential and Senate candidates, multiple two-year cycles exist between elections.
}
