% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fecapi-package.R
\docType{package}
\name{fecapi-package}
\alias{fecapi}
\alias{fecapi-package}
\title{fecapi: 'OpenFEC' 'API' Client}
\description{
A client to interact with the 'OpenFEC' 'API' at \url{https://api.open.fec.gov/developers/}. The 'OpenFEC' 'API' provides programmatic access to campaign finance data stored by the Federal Election Commission.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonthegeek/fecapi}
  \item \url{https://jonthegeek.github.io/fecapi/}
  \item Report bugs at \url{https://github.com/jonthegeek/fecapi/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
